requirements.txt
discord
python-dotenv
yt-dlp
pillow
pynacl
google-api-python-client
ffmpeg
jishaku
aiofiles